added and tested table resources + modified guest resource

+++ IS3106WeddingPlanner-ejb/src/java/session/GuestTableSessionBean.java
@@ -8,7 +8,10 @@ package session;
 import entity.Guest;
 import entity.GuestTable;
 import entity.WeddingProject;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -17,6 +20,8 @@ import javax.persistence.PersistenceContext;
  * @author leomi
  */
 import util.exception.InvalidAssociationException;
+import util.exception.InvalidDeleteException;
+import util.exception.InvalidGetException;
 @Stateless
 public class GuestTableSessionBean implements GuestTableSessionBeanLocal {
 
@@ -24,46 +29,76 @@ public class GuestTableSessionBean implements GuestTableSessionBeanLocal {
     private EntityManager em;
 
     @Override
-    public void createGuestTable(GuestTable guestTable, Long weddingProjectId) {
+    public void createGuestTable(GuestTable guestTable, Long weddingProjectId) throws InvalidAssociationException {
         WeddingProject weddingProject = em.find(WeddingProject.class, weddingProjectId);
-        em.persist(guestTable);
-        guestTable.setWeddingProject(weddingProject);
-        weddingProject.getTables().add(guestTable);
+        if (weddingProject != null) {
+            em.persist(guestTable);
+            guestTable.setWeddingProject(weddingProject);
+            weddingProject.getTables().add(guestTable);
+        } else {
+            throw new InvalidAssociationException();
+        }
     }
     
     @Override
     public void addGuestToTable(Long guestId, Long guestTableId) throws InvalidAssociationException {
         Guest guest = em.find(Guest.class, guestId);
         GuestTable guestTable = em.find(GuestTable.class, guestTableId);
-        if (!guest.getWeddingProject().equals(guestTable.getWeddingProject())) {
+        if (guest != null && guestTable != null) {
+            if (!guest.getWeddingProject().equals(guestTable.getWeddingProject())) {
+                throw new InvalidAssociationException();
+            }
+            guest.setGuestTable(guestTable);
+            guestTable.getGuests().add(guest);
+        } else {
             throw new InvalidAssociationException();
         }
-        guest.setGuestTable(guestTable);
-        guestTable.getGuests().add(guest);
     }
     
     @Override
     public void removeGuestFromTable(Long guestId, Long guestTableId) throws InvalidAssociationException {
         Guest guest = em.find(Guest.class, guestId);
         GuestTable guestTable = em.find(GuestTable.class, guestTableId);
-        if (!guest.getWeddingProject().equals(guestTable.getWeddingProject())) {
+        if (guest != null && guestTable != null) {
+            if (!guest.getWeddingProject().equals(guestTable.getWeddingProject())) {
+                throw new InvalidAssociationException();
+            }
+            guest.setGuestTable(null);
+            guestTable.getGuests().remove(guest);
+        } else {
             throw new InvalidAssociationException();
         }
-        guest.setGuestTable(null);
-        guestTable.getGuests().remove(guest);
     }
     
     @Override
-    public void deleteGuestTable(Long guestTableId) {
+    public void deleteGuestTable(Long guestTableId) throws InvalidDeleteException {
         GuestTable guestTable = em.find(GuestTable.class, guestTableId);
-        WeddingProject weddingProject = guestTable.getWeddingProject();
-        if (!guestTable.getGuests().isEmpty()) {
-            List<Guest> guests = guestTable.getGuests();
-            guests.forEach(guest -> guest.setGuestTable(null));
-            guestTable.getGuests().clear();
+        if (guestTable != null) {
+            WeddingProject weddingProject = guestTable.getWeddingProject();
+            if (!guestTable.getGuests().isEmpty()) {
+                List<Guest> guests = guestTable.getGuests();
+                guests.forEach(guest -> guest.setGuestTable(null));
+                guestTable.getGuests().clear();
+            }
+            weddingProject.getTables().remove(guestTable);
+            em.remove(guestTable);
+        } else {
+            throw new InvalidDeleteException();
         }
-        weddingProject.getTables().remove(guestTable);
-        em.remove(guestTable);
+    }
+    
+    @Override
+    public List<GuestTable> getGuestTables(Long weddingId) throws InvalidGetException {
+        if (em.find(WeddingProject.class, weddingId) != null) {
+            Stream <GuestTable> tables = em.createQuery("SELECT g FROM GuestTable g WHERE g.weddingProject.weddingProjectId = ?1").setParameter(1, weddingId)
+                                            .getResultStream();
+            return tables.map(g -> {
+                                em.detach(g);
+                                return g;
+                                }).collect(Collectors.toList());
+        } else {
+            throw new InvalidGetException();
+        }
     }
 
     // Add business logic below. (Right-click in editor and choose
